backend/requirements.txt
fastapi==0.115.5
uvicorn[standard]==0.32.1
redis==5.2.1
openai==1.57.0
anthropic==0.40.0
langgraph==0.2.60
langchain-core==0.2.43
pydantic==2.10.3
faiss-cpu==1.9.0.post1
sentence-transformers==3.3.1
tavily-python==0.5.0
httpx==0.28.1
python-dotenv==1.0.1
websockets==14.1
